+++ Code/AT_1Init.lua
@@ -17,6 +17,7 @@ g_AT_Options = {
 	ATvoyageWaitTime    = 5,         -- Wait this amount of sols between voyages
 	ATrecallRadius      = true,      -- display recall radius on landed rocket
 	ATearlyDepartures   = true,      -- allow for earlier departures when voyages waiting
+	ATstripSpecialty    = true,      -- strip a tourists specialty upon arrival
 } -- g_AT_Options
 
 -- Save game fixup variables
@@ -275,14 +276,16 @@ function OnMsg.RocketLaunchFromEarth(rocket)
       local cargo = {}
 
       if #tourists > 0 then
-      	-- remove specializations from tourists
-      	for i = 1, #tourists do
-      		if tourists[i].specialist ~= "none" then
-      			tourists[i].traits[tourists[i].specialist] = nil
-      			tourists[i].specialist = "none"
-      			tourists[i].traits.none = true
-      		end -- if tourists[i]
-      	end -- for i
+      	if g_AT_Options.ATstripSpecialty then
+      	  -- remove specializations from tourists
+      	  for i = 1, #tourists do
+      	  	if tourists[i].specialist ~= "none" then
+      	  		tourists[i].traits[tourists[i].specialist] = nil
+      	  		tourists[i].specialist = "none"
+      	  		tourists[i].traits.none = true
+      	  	end -- if tourists[i]
+      	  end -- for i
+      	end -- if g_AT_Options.ATstripSpecialty
 
         cargo[1] = {
           class = "Passengers",